Total revenue for merchant

+++ app/models/item.rb
@@ -1,5 +1,8 @@
 class Item < ApplicationRecord
-  validates_presence_of(:name, :description, :unit_price, :merchant_id)
+  validates :name, presence: true
+  validates :description, presence: true
+  validates :unit_price, presence: true
+  validates :merchant_id, presence: true
   validates :unit_price, numericality: { greater_than_or_equal_to: 0 }, presence: true
 
   # this must come before has_many invoice_items https://stackoverflow.com/a/31344318
@@ -9,14 +12,11 @@ class Item < ApplicationRecord
   has_many :invoice_items, dependent: :destroy
   has_many :invoices, through: :invoice_items
 
-
   private
 
   def remove_empty_invoices
-    self.invoices.each do |invoice|
-      if invoice.items.count == 1
-        invoice.destroy
-      end
+    invoices.each do |invoice|
+      invoice.destroy if invoice.items.count == 1
     end
   end
 end
